curve typescript
